change animal

+++ AnimalNursery/Models/Base/Animal.cs
@@ -7,7 +7,7 @@ namespace AnimalNursery.Models.Base
     {
         private int _id;
         private string _name;
-        private int _age;
+        private double _age;
         private double _height;
         private double _weight;
         private string _classAnimal;
@@ -15,13 +15,13 @@ namespace AnimalNursery.Models.Base
 
         public int Id { get => _id; set { _id = value; OnPropertyChanged(nameof(Id)); } }
         public string Name { get => _name; set { _name = value; OnPropertyChanged(nameof(Name)); } }
-        public int Age { get => _age; set { _age = value; OnPropertyChanged(nameof(Age)); } }
+        public double Age { get => _age; set { _age = value; OnPropertyChanged(nameof(Age)); } }
         public double Height { get => _height; set { _height = value; OnPropertyChanged(nameof(Height)); } }
         public double Weight { get => _weight; set { _weight = value; OnPropertyChanged(nameof(Weight)); } }
         public string ClassAnimal { get => _classAnimal; set { _classAnimal = value; OnPropertyChanged(nameof(ClassAnimal)); } }
         public string Command { get => _command; set { _command = value; OnPropertyChanged(nameof(Command)); } }
 
-        public Animal(int id, string name, int age, double height, double weight, string classAnimal, string command)
+        public Animal(int id, string name, double age, double height, double weight, string classAnimal, string command)
         {
             Id = id;
             Name = name;
